+++ docs/source/ingest/destination_connectors/code/bash/sql.sh
@@ -9,7 +9,7 @@ unstructured-ingest \
   --verbose \
   --strategy fast \
   sql \
-  --db_name postgresql \
+  --db_type postgresql \
   --username postgres \
   --password test \
   --host localhost \