+++ test_unstructured_ingest/dest/pgvector.sh
@@ -11,13 +11,13 @@ OUTPUT_DIR=$OUTPUT_ROOT/structured-output/$OUTPUT_FOLDER_NAME
 WORK_DIR=$OUTPUT_ROOT/workdir/$OUTPUT_FOLDER_NAME
 max_processes=${MAX_PROCESSES:=$(python3 -c "import os; print(os.cpu_count())")}
 CI=${CI:-"false"}
-DATABASE_NAME="pgvector"
+DATABASE_TYPE="pgvector"
 
 # shellcheck disable=SC1091
 source "$SCRIPT_DIR"/cleanup.sh
 function cleanup {
   echo "Stopping SQL DB Docker container"
-  docker-compose -f scripts/sql-test-helpers/docker-compose-"$DATABASE_NAME".yaml down --remove-orphans -v
+  docker-compose -f scripts/sql-test-helpers/docker-compose-"$DATABASE_TYPE".yaml down --remove-orphans -v
   # Local file cleanup
   cleanup_dir "$WORK_DIR"
   cleanup_dir "$OUTPUT_DIR"
@@ -31,7 +31,7 @@ trap cleanup EXIT
 # Create sql instance and create `elements` class
 echo "Creating SQL DB instance"
 # shellcheck source=/dev/null
-scripts/sql-test-helpers/create-sql-instance.sh "$DATABASE_NAME"
+scripts/sql-test-helpers/create-sql-instance.sh "$DATABASE_TYPE"
 wait
 
 PYTHONPATH=. ./unstructured/ingest/main.py \
@@ -44,11 +44,11 @@ PYTHONPATH=. ./unstructured/ingest/main.py \
   --work-dir "$WORK_DIR" \
   --embedding-provider "langchain-huggingface" \
   sql \
-  --db_name "postgresql" \
+  --db_type "postgresql" \
   --username unstructured \
   --password test \
   --host localhost \
   --port 5433 \
   --database elements
 
-"$SCRIPT_DIR"/python/test-ingest-sql-output.py "$DATABASE_NAME" "5433"
+"$SCRIPT_DIR"/python/test-ingest-sql-output.py "$DATABASE_TYPE" "5433"